+++ src/assets/styles/HomeStyle.js
@@ -1,0 +1,6 @@
+import styled from "styled-components";
+import { colors } from "../utils/colors";
+
+export const Title = styled.p`
+  color: ${colors.pink};
+`;
