--- src/assets/styles/home.js
@@ -1,8 +1,0 @@
-import styled from "styled-components";
-import { colors } from "../../utils/colors";
-import { fonts } from '../../utils/fonts'
-
-export const Title = styled.p`
-  color: ${colors.pink};
-  font-family: '${fonts.light}';
-`
