+++ src/assets/utils/colors.js
@@ -1,0 +1,6 @@
+export const colors = {
+    black: '#FFFFFF',
+    white: '#000000',
+    pink: '#F14866',
+    gray: '#E6E1DD',
+};
