--- src/utils/colors.js
@@ -1,10 +1,0 @@
-export const colors = {
-    white: '#FFFFFF',
-    black: '#000000',
-    pink: '#F14866',
-    pinkHover: '#F0395A',
-    gray: '#E6E1DD',
-    blackGray: '#555555',
-    grayMedium: '#969696',
-    red: '#FF5757',
-};
