--- src/utils/fonts.js
@@ -1,8 +1,0 @@
-export const fonts = {
-    light: "PoppinsLight",
-    thin: "PoppinsThin",
-    regular: "PoppinsRegular",
-    medium: "PoppinsMedium",
-    semiBold: "PoppinsSemiBold",
-    bold: "PoppinsBold",
-}
